Time: 118 ms (13.54%) | Memory: 55.9 MB (68.75%) - LeetSync

+++ 3107-maximum-spending-after-buying-items/maximum-spending-after-buying-items.java
@@ -13,7 +13,7 @@ class Solution {
 
         int k = 1;
         while (!pq.isEmpty()) {
-            spending += (double) pq.poll() * k++;
+            spending += (long) pq.poll() * k++;
         }
 
         return spending;
